// lib/api/schema.ts
export interface components {
    schemas: {
// ...
        QueueEntriesPublic: {
            /** Data */
            data: components["schemas"]["QueueEntryPublic"][];
            /** Count */
            count: number;
        };
        /** QueueEntryPublic */
        QueueEntryPublic: {
            /** Ticket Number */
            ticket_number: number;
            /**
             * Status
             * @default waiting
             */
            status: string;
            /** Source */
            source: string;
            /** Guest Name */
            guest_name?: string | null;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /**
             * Service Item Id
             * Format: uuid
             */
            service_item_id: string;
            /** User Id */
            user_id?: string | null;
            /** Joined At */
            joined_at?: string | null;
            /** Completed At */
            completed_at?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}